requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.2
httpx>=0.25.2
aiohttp>=3.9.1
click>=8.1.7
rich>=13.7.0
jinja2>=3.1.2
pyyaml>=6.0.1
python-multipart>=0.0.6
aiofiles>=23.2.1
PyGithub>=2.1.1
pynacl>=1.5.0
